refactor: UI컴포넌트분리
